Reorganize data access layer files
With only one repository to track, I'm going to keep things simple and keep all data access layer files, including model and repository classes, in a single namespace and folder. We can always split things up later if needed.
