+++ FireMothServices.Tests/DataAccess/FileFingerprintTests.cs
@@ -9,7 +9,7 @@ namespace RiotClub.FireMoth.Services.Tests.FileScanning
     using System.Collections.Generic;
     using System.Diagnostics.CodeAnalysis;
     using System.IO.Abstractions.TestingHelpers;
-    using RiotClub.FireMoth.Services.DataAccess.Model;
+    using RiotClub.FireMoth.Services.DataAccess;
     using RiotClub.FireMoth.Services.DataAnalysis;
     using Xunit;
